requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0 